omni-music-pro/pyproject.toml
[tool.poetry]
name = "omni-music-pro"
version = "1.0.0"
description = "Application moderne de téléchargement et gestion musicale"
authors = ["Greg <greg@omnirealm.com>"]
readme = "README.md"
license = "MIT"
packages = [{include = "omni_music", from = "src"}]

[tool.poetry.dependencies]
python = "^3.12"
customtkinter = "^5.2.0"
yt-dlp = "^2025.8.11"
mutagen = "^1.47.0"
aiohttp = "^3.9.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
pillow = "^10.0.0"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^24.0.0"
ruff = "^0.1.0"
mypy = "^1.8.0"
pre-commit = "^3.6.0"
safety = "^3.6.0"

[tool.poetry.scripts]
omni-music = "omni_music.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "N", "W", "B", "C90", "UP", "ARG", "SIM", "PL"]
ignore = ["E501"]  # Line too long, handled by black

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=omni_music",
    "--cov-branch",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["src/omni_music"]
omit = [
    "*/tests/*", 
    "*/__init__.py",
    "*/presentation/gui.py",  # Interface graphique exclue des tests
    "*/__main__.py"           # Point d'entrée exclu des tests
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = true